Added Authorization Checks

+++ controllers/feed.js
@@ -128,6 +128,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if(post.creator.toString() !== req.userId){
+        const error = new Error("Not Authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -157,6 +164,13 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if(post.creator.toString() !== req.userId){
+        const error = new Error("Not Authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
 
       return Post.findByIdAndDelete(postId);
